Set up graph file to run simulations upon

Set up graph file to run simulations upon

disnet_sim_setup(g, r0 = 1.44, latent_period = 2.62,
  inf_period = 3.38, tau = 3, r_beta = 0.5, p_a = 1/3,
  seed_nd = NA, seed_no = 1, vacc = FALSE,
  output_dir = getOption("disnetOutputDir", NA), vacc_eff = c(80, 90),
  vacc_cov = c(80, 100), vacc_nd = "890", ...)

Arguments

g
The network object created after commuting rates were calculated

r0
Reproduction number

latent_period
Period between infection and becoming infectious

inf_period
Duration of infectious period

tau
Return rate (from commuting)

r_beta
Reduction in infectiousness due to being aymptomatic inf

p_a
Probability of asymptomatic infection

seed_nd
The node which you want to seed/initiate infection in

seed_no
The number of infected people you want to start inf with

vacc
Whether a vaccination campaign is underway or not

output_dir
Directory to store intermediate simulation data in.

vacc_eff
Vaccination efficacy

vacc_cov
Vaccination coverage

vacc_nd
Node you want to vaccinate in

Details

1. Reads in an `igraph` object (with commuting proportions previously calculated)

2. Calculates and attaches "Effective Population" to graph

3. Vaccinates target nodes with vaccination proportion: vaccination proportions = vaccination coverage * vaccination efficacy The proportion of a node that is vaccinated is removed from the Susceptible compartment and added to the Recovered compartment

3. Seeds a node with specified number of infectious individuals (taking them out from the Susceptibles and adding to Infectious compartment

4. Creates the FOI structure/scaffold used to quickly calculate FOI (FOI = Force of Infection, acting on an individual in a node)

Examples

f = system.file("sampleData", "g.rds", package = "disnet")
g = readRDS(f)
g_comm = disnet_commuting(g)

#> [1] "working on node: 7"
#> [1] "working on node: 10"
#> [1] "working on node: 20"
#> [1] "working on node: 33"
#> [1] "working on node: 38"
#> [1] "working on node: 58"
#> [1] "working on node: 81"
#> [1] "working on node: 97"
#> [1] "working on node: 110"
#> [1] "working on node: 112"
#> [1] "working on node: 134"
#> [1] "working on node: 136"
#> [1] "working on node: 138"
#> [1] "working on node: 140"
#> [1] "working on node: 150"
#> [1] "working on node: 152"
#> [1] "working on node: 163"
#> [1] "working on node: 182"
#> [1] "working on node: 185"
#> [1] "working on node: 197"
#> [1] "working on node: 231"
#> [1] "working on node: 241"
#> [1] "working on node: 245"
#> [1] "working on node: 257"
#> [1] "working on node: 261"
#> [1] "working on node: 265"
#> [1] "working on node: 273"
#> [1] "working on node: 274"
#> [1] "working on node: 318"
#> [1] "working on node: 321"
#> [1] "working on node: 337"
#> [1] "working on node: 363"
#> [1] "working on node: 382"
#> [1] "working on node: 412"
#> [1] "working on node: 415"
#> [1] "working on node: 416"
#> [1] "working on node: 427"
#> [1] "working on node: 446"
#> [1] "working on node: 477"
#> [1] "working on node: 501"
#> [1] "working on node: 539"
#> [1] "working on node: 554"
#> [1] "working on node: 570"
#> [1] "working on node: 615"
#> [1] "working on node: 626"
#> [1] "working on node: 627"
#> [1] "working on node: 642"
#> [1] "working on node: 643"
#> [1] "working on node: 653"
#> [1] "working on node: 666"
#> [1] "working on node: 667"
#> [1] "working on node: 704"
#> [1] "working on node: 738"
#> [1] "working on node: 739"
#> [1] "working on node: 742"
#> [1] "working on node: 746"
#> [1] "working on node: 754"
#> [1] "working on node: 760"
#> [1] "working on node: 761"
#> [1] "working on node: 765"
#> [1] "working on node: 771"
#> [1] "working on node: 773"
#> [1] "working on node: 774"
#> [1] "working on node: 787"
#> [1] "working on node: 792"
#> [1] "working on node: 819"
#> [1] "working on node: 820"
#> [1] "working on node: 824"
#> [1] "working on node: 842"
#> [1] "working on node: 846"
#> [1] "working on node: 858"
#> [1] "working on node: 865"
#> [1] "working on node: 867"
#> [1] "working on node: 872"
#> [1] "working on node: 877"
#> [1] "working on node: 884"
#> [1] "working on node: 906"
#> [1] "working on node: 908"
#> [1] "working on node: 911"
#> [1] "working on node: 930"
#> [1] "working on node: 944"
#> [1] "working on node: 977"
#> [1] "working on node: 978"

nodes = igraph::vcount(g_comm)
set.seed(890)
seed_nd = igraph::vertex_attr(g_comm, "name", sample(1:nodes, 1))
for_sim = disnet_sim_setup(g_comm, seed_nd = seed_nd, output_dir = NA)