Main simulation function

This function wraps up all simulation sub-functins and applies them to the intermediate sim files (created by `disnet_sim_setup`

disnet_simulate(nsims = 10, nsteps = 1000, sim_input = sim_intermed,
  sim_output_dir = getOption("disnetOutputDir", NA), parallel = FALSE,
  seed = 0)

Arguments

nsims
The number of simulations to run. Default is 10

nsteps
The number of timesteps you'd like each simulation to run for. Default = 1000. It's set to a large value so that outbreaks can run their course.

sim_input
The file setup/created by disnet_sim_setup, which has incorprated the foi, infection seeding, and/or vaccination campaign.

sim_output_dir
The directory you'd like to save the simulation results to.

Examples

f = system.file("sampleData", "g.rds", package = "disnet")
g = readRDS(f)
g_comm = disnet_commuting(g)

#> [1] "working on node: 7"
#> [1] "working on node: 10"
#> [1] "working on node: 20"
#> [1] "working on node: 33"
#> [1] "working on node: 38"
#> [1] "working on node: 58"
#> [1] "working on node: 81"
#> [1] "working on node: 97"
#> [1] "working on node: 110"
#> [1] "working on node: 112"
#> [1] "working on node: 134"
#> [1] "working on node: 136"
#> [1] "working on node: 138"
#> [1] "working on node: 140"
#> [1] "working on node: 150"
#> [1] "working on node: 152"
#> [1] "working on node: 163"
#> [1] "working on node: 182"
#> [1] "working on node: 185"
#> [1] "working on node: 197"
#> [1] "working on node: 231"
#> [1] "working on node: 241"
#> [1] "working on node: 245"
#> [1] "working on node: 257"
#> [1] "working on node: 261"
#> [1] "working on node: 265"
#> [1] "working on node: 273"
#> [1] "working on node: 274"
#> [1] "working on node: 318"
#> [1] "working on node: 321"
#> [1] "working on node: 337"
#> [1] "working on node: 363"
#> [1] "working on node: 382"
#> [1] "working on node: 412"
#> [1] "working on node: 415"
#> [1] "working on node: 416"
#> [1] "working on node: 427"
#> [1] "working on node: 446"
#> [1] "working on node: 477"
#> [1] "working on node: 501"
#> [1] "working on node: 539"
#> [1] "working on node: 554"
#> [1] "working on node: 570"
#> [1] "working on node: 615"
#> [1] "working on node: 626"
#> [1] "working on node: 627"
#> [1] "working on node: 642"
#> [1] "working on node: 643"
#> [1] "working on node: 653"
#> [1] "working on node: 666"
#> [1] "working on node: 667"
#> [1] "working on node: 704"
#> [1] "working on node: 738"
#> [1] "working on node: 739"
#> [1] "working on node: 742"
#> [1] "working on node: 746"
#> [1] "working on node: 754"
#> [1] "working on node: 760"
#> [1] "working on node: 761"
#> [1] "working on node: 765"
#> [1] "working on node: 771"
#> [1] "working on node: 773"
#> [1] "working on node: 774"
#> [1] "working on node: 787"
#> [1] "working on node: 792"
#> [1] "working on node: 819"
#> [1] "working on node: 820"
#> [1] "working on node: 824"
#> [1] "working on node: 842"
#> [1] "working on node: 846"
#> [1] "working on node: 858"
#> [1] "working on node: 865"
#> [1] "working on node: 867"
#> [1] "working on node: 872"
#> [1] "working on node: 877"
#> [1] "working on node: 884"
#> [1] "working on node: 906"
#> [1] "working on node: 908"
#> [1] "working on node: 911"
#> [1] "working on node: 930"
#> [1] "working on node: 944"
#> [1] "working on node: 977"
#> [1] "working on node: 978"

nodes = igraph::vcount(g_comm)
set.seed(890)
seed_nd = igraph::vertex_attr(g_comm, "name", sample(1:nodes, 1))
for_sim = disnet_sim_setup(g_comm, seed_nd = seed_nd, output_dir = NA)
simres = disnet_simulate(sim_input = for_sim)

#> 
#> Starting simulations

#> -------------------------------------------------------
#> ******** Simulation  1 
#> 			Timestep: 1/1000			Timestep: 2/1000			Timestep: 3/1000			Timestep: 4/1000			Timestep: 5/1000			Timestep: 6/1000-------------------------------------------------------
#> ******** Simulation  2 
#> 			Timestep: 1/1000			Timestep: 2/1000			Timestep: 3/1000			Timestep: 4/1000			Timestep: 5/1000			Timestep: 6/1000			Timestep: 7/1000			Timestep: 8/1000			Timestep: 9/1000			Timestep: 10/1000			Timestep: 11/1000			Timestep: 12/1000			Timestep: 13/1000			Timestep: 14/1000			Timestep: 15/1000			Timestep: 16/1000			Timestep: 17/1000			Timestep: 18/1000			Timestep: 19/1000			Timestep: 20/1000			Timestep: 21/1000			Timestep: 22/1000			Timestep: 23/1000			Timestep: 24/1000			Timestep: 25/1000			Timestep: 26/1000			Timestep: 27/1000			Timestep: 28/1000			Timestep: 29/1000			Timestep: 30/1000			Timestep: 31/1000			Timestep: 32/1000			Timestep: 33/1000			Timestep: 34/1000			Timestep: 35/1000			Timestep: 36/1000			Timestep: 37/1000			Timestep: 38/1000-------------------------------------------------------
#> ******** Simulation  3 
#> 			Timestep: 1/1000			Timestep: 2/1000			Timestep: 3/1000			Timestep: 4/1000			Timestep: 5/1000			Timestep: 6/1000			Timestep: 7/1000			Timestep: 8/1000			Timestep: 9/1000-------------------------------------------------------
#> ******** Simulation  4 
#> 			Timestep: 1/1000			Timestep: 2/1000			Timestep: 3/1000			Timestep: 4/1000			Timestep: 5/1000			Timestep: 6/1000			Timestep: 7/1000			Timestep: 8/1000			Timestep: 9/1000			Timestep: 10/1000			Timestep: 11/1000			Timestep: 12/1000			Timestep: 13/1000			Timestep: 14/1000			Timestep: 15/1000			Timestep: 16/1000			Timestep: 17/1000			Timestep: 18/1000			Timestep: 19/1000			Timestep: 20/1000			Timestep: 21/1000			Timestep: 22/1000			Timestep: 23/1000			Timestep: 24/1000			Timestep: 25/1000			Timestep: 26/1000			Timestep: 27/1000			Timestep: 28/1000			Timestep: 29/1000			Timestep: 30/1000			Timestep: 31/1000			Timestep: 32/1000			Timestep: 33/1000			Timestep: 34/1000			Timestep: 35/1000			Timestep: 36/1000			Timestep: 37/1000			Timestep: 38/1000			Timestep: 39/1000			Timestep: 40/1000			Timestep: 41/1000			Timestep: 42/1000			Timestep: 43/1000			Timestep: 44/1000			Timestep: 45/1000			Timestep: 46/1000			Timestep: 47/1000			Timestep: 48/1000			Timestep: 49/1000			Timestep: 50/1000			Timestep: 51/1000			Timestep: 52/1000			Timestep: 53/1000			Timestep: 54/1000			Timestep: 55/1000			Timestep: 56/1000			Timestep: 57/1000			Timestep: 58/1000			Timestep: 59/1000			Timestep: 60/1000			Timestep: 61/1000			Timestep: 62/1000			Timestep: 63/1000			Timestep: 64/1000			Timestep: 65/1000			Timestep: 66/1000			Timestep: 67/1000			Timestep: 68/1000			Timestep: 69/1000			Timestep: 70/1000			Timestep: 71/1000			Timestep: 72/1000			Timestep: 73/1000			Timestep: 74/1000			Timestep: 75/1000			Timestep: 76/1000			Timestep: 77/1000			Timestep: 78/1000			Timestep: 79/1000			Timestep: 80/1000			Timestep: 81/1000			Timestep: 82/1000			Timestep: 83/1000			Timestep: 84/1000			Timestep: 85/1000-------------------------------------------------------
#> ******** Simulation  5 
#> 			Timestep: 1/1000			Timestep: 2/1000			Timestep: 3/1000			Timestep: 4/1000			Timestep: 5/1000			Timestep: 6/1000			Timestep: 7/1000			Timestep: 8/1000			Timestep: 9/1000			Timestep: 10/1000			Timestep: 11/1000			Timestep: 12/1000			Timestep: 13/1000			Timestep: 14/1000			Timestep: 15/1000			Timestep: 16/1000			Timestep: 17/1000			Timestep: 18/1000			Timestep: 19/1000			Timestep: 20/1000			Timestep: 21/1000			Timestep: 22/1000			Timestep: 23/1000			Timestep: 24/1000			Timestep: 25/1000			Timestep: 26/1000			Timestep: 27/1000			Timestep: 28/1000			Timestep: 29/1000			Timestep: 30/1000			Timestep: 31/1000			Timestep: 32/1000			Timestep: 33/1000			Timestep: 34/1000			Timestep: 35/1000			Timestep: 36/1000			Timestep: 37/1000			Timestep: 38/1000			Timestep: 39/1000			Timestep: 40/1000			Timestep: 41/1000			Timestep: 42/1000			Timestep: 43/1000			Timestep: 44/1000			Timestep: 45/1000			Timestep: 46/1000			Timestep: 47/1000			Timestep: 48/1000			Timestep: 49/1000			Timestep: 50/1000			Timestep: 51/1000			Timestep: 52/1000			Timestep: 53/1000			Timestep: 54/1000			Timestep: 55/1000			Timestep: 56/1000			Timestep: 57/1000			Timestep: 58/1000			Timestep: 59/1000			Timestep: 60/1000			Timestep: 61/1000			Timestep: 62/1000			Timestep: 63/1000			Timestep: 64/1000			Timestep: 65/1000			Timestep: 66/1000			Timestep: 67/1000			Timestep: 68/1000			Timestep: 69/1000			Timestep: 70/1000			Timestep: 71/1000			Timestep: 72/1000			Timestep: 73/1000			Timestep: 74/1000			Timestep: 75/1000			Timestep: 76/1000			Timestep: 77/1000			Timestep: 78/1000			Timestep: 79/1000			Timestep: 80/1000			Timestep: 81/1000			Timestep: 82/1000			Timestep: 83/1000			Timestep: 84/1000			Timestep: 85/1000			Timestep: 86/1000			Timestep: 87/1000			Timestep: 88/1000			Timestep: 89/1000			Timestep: 90/1000			Timestep: 91/1000			Timestep: 92/1000			Timestep: 93/1000			Timestep: 94/1000			Timestep: 95/1000			Timestep: 96/1000			Timestep: 97/1000			Timestep: 98/1000			Timestep: 99/1000			Timestep: 100/1000			Timestep: 101/1000			Timestep: 102/1000			Timestep: 103/1000			Timestep: 104/1000			Timestep: 105/1000			Timestep: 106/1000			Timestep: 107/1000			Timestep: 108/1000			Timestep: 109/1000			Timestep: 110/1000			Timestep: 111/1000			Timestep: 112/1000			Timestep: 113/1000			Timestep: 114/1000			Timestep: 115/1000			Timestep: 116/1000			Timestep: 117/1000			Timestep: 118/1000			Timestep: 119/1000			Timestep: 120/1000			Timestep: 121/1000			Timestep: 122/1000			Timestep: 123/1000			Timestep: 124/1000			Timestep: 125/1000			Timestep: 126/1000			Timestep: 127/1000			Timestep: 128/1000			Timestep: 129/1000			Timestep: 130/1000			Timestep: 131/1000			Timestep: 132/1000			Timestep: 133/1000			Timestep: 134/1000			Timestep: 135/1000			Timestep: 136/1000			Timestep: 137/1000			Timestep: 138/1000			Timestep: 139/1000			Timestep: 140/1000			Timestep: 141/1000			Timestep: 142/1000			Timestep: 143/1000			Timestep: 144/1000			Timestep: 145/1000			Timestep: 146/1000			Timestep: 147/1000			Timestep: 148/1000			Timestep: 149/1000			Timestep: 150/1000			Timestep: 151/1000			Timestep: 152/1000			Timestep: 153/1000			Timestep: 154/1000			Timestep: 155/1000			Timestep: 156/1000			Timestep: 157/1000			Timestep: 158/1000			Timestep: 159/1000			Timestep: 160/1000			Timestep: 161/1000			Timestep: 162/1000			Timestep: 163/1000			Timestep: 164/1000			Timestep: 165/1000			Timestep: 166/1000			Timestep: 167/1000			Timestep: 168/1000			Timestep: 169/1000			Timestep: 170/1000			Timestep: 171/1000			Timestep: 172/1000			Timestep: 173/1000			Timestep: 174/1000			Timestep: 175/1000			Timestep: 176/1000			Timestep: 177/1000			Timestep: 178/1000			Timestep: 179/1000			Timestep: 180/1000			Timestep: 181/1000			Timestep: 182/1000			Timestep: 183/1000			Timestep: 184/1000			Timestep: 185/1000			Timestep: 186/1000			Timestep: 187/1000			Timestep: 188/1000			Timestep: 189/1000			Timestep: 190/1000			Timestep: 191/1000			Timestep: 192/1000			Timestep: 193/1000			Timestep: 194/1000			Timestep: 195/1000			Timestep: 196/1000			Timestep: 197/1000			Timestep: 198/1000			Timestep: 199/1000			Timestep: 200/1000			Timestep: 201/1000			Timestep: 202/1000			Timestep: 203/1000			Timestep: 204/1000			Timestep: 205/1000			Timestep: 206/1000			Timestep: 207/1000			Timestep: 208/1000			Timestep: 209/1000			Timestep: 210/1000			Timestep: 211/1000			Timestep: 212/1000			Timestep: 213/1000			Timestep: 214/1000			Timestep: 215/1000			Timestep: 216/1000			Timestep: 217/1000			Timestep: 218/1000			Timestep: 219/1000			Timestep: 220/1000			Timestep: 221/1000			Timestep: 222/1000			Timestep: 223/1000			Timestep: 224/1000			Timestep: 225/1000			Timestep: 226/1000			Timestep: 227/1000			Timestep: 228/1000			Timestep: 229/1000			Timestep: 230/1000			Timestep: 231/1000			Timestep: 232/1000			Timestep: 233/1000			Timestep: 234/1000			Timestep: 235/1000			Timestep: 236/1000			Timestep: 237/1000			Timestep: 238/1000			Timestep: 239/1000			Timestep: 240/1000			Timestep: 241/1000			Timestep: 242/1000			Timestep: 243/1000			Timestep: 244/1000			Timestep: 245/1000			Timestep: 246/1000			Timestep: 247/1000			Timestep: 248/1000			Timestep: 249/1000			Timestep: 250/1000			Timestep: 251/1000			Timestep: 252/1000			Timestep: 253/1000			Timestep: 254/1000			Timestep: 255/1000			Timestep: 256/1000			Timestep: 257/1000			Timestep: 258/1000			Timestep: 259/1000			Timestep: 260/1000			Timestep: 261/1000			Timestep: 262/1000			Timestep: 263/1000			Timestep: 264/1000			Timestep: 265/1000			Timestep: 266/1000			Timestep: 267/1000			Timestep: 268/1000			Timestep: 269/1000			Timestep: 270/1000			Timestep: 271/1000			Timestep: 272/1000			Timestep: 273/1000			Timestep: 274/1000			Timestep: 275/1000			Timestep: 276/1000			Timestep: 277/1000			Timestep: 278/1000			Timestep: 279/1000-------------------------------------------------------
#> ******** Simulation  6 
#> 			Timestep: 1/1000			Timestep: 2/1000			Timestep: 3/1000			Timestep: 4/1000			Timestep: 5/1000			Timestep: 6/1000			Timestep: 7/1000			Timestep: 8/1000			Timestep: 9/1000			Timestep: 10/1000			Timestep: 11/1000			Timestep: 12/1000			Timestep: 13/1000			Timestep: 14/1000			Timestep: 15/1000			Timestep: 16/1000			Timestep: 17/1000			Timestep: 18/1000			Timestep: 19/1000			Timestep: 20/1000			Timestep: 21/1000			Timestep: 22/1000			Timestep: 23/1000			Timestep: 24/1000			Timestep: 25/1000			Timestep: 26/1000			Timestep: 27/1000			Timestep: 28/1000			Timestep: 29/1000			Timestep: 30/1000			Timestep: 31/1000			Timestep: 32/1000			Timestep: 33/1000			Timestep: 34/1000			Timestep: 35/1000			Timestep: 36/1000			Timestep: 37/1000			Timestep: 38/1000			Timestep: 39/1000			Timestep: 40/1000			Timestep: 41/1000			Timestep: 42/1000			Timestep: 43/1000			Timestep: 44/1000			Timestep: 45/1000			Timestep: 46/1000			Timestep: 47/1000			Timestep: 48/1000			Timestep: 49/1000			Timestep: 50/1000			Timestep: 51/1000			Timestep: 52/1000			Timestep: 53/1000			Timestep: 54/1000			Timestep: 55/1000			Timestep: 56/1000			Timestep: 57/1000			Timestep: 58/1000			Timestep: 59/1000			Timestep: 60/1000			Timestep: 61/1000			Timestep: 62/1000			Timestep: 63/1000			Timestep: 64/1000			Timestep: 65/1000			Timestep: 66/1000			Timestep: 67/1000			Timestep: 68/1000			Timestep: 69/1000			Timestep: 70/1000			Timestep: 71/1000			Timestep: 72/1000			Timestep: 73/1000			Timestep: 74/1000			Timestep: 75/1000			Timestep: 76/1000			Timestep: 77/1000			Timestep: 78/1000			Timestep: 79/1000			Timestep: 80/1000			Timestep: 81/1000			Timestep: 82/1000			Timestep: 83/1000			Timestep: 84/1000			Timestep: 85/1000			Timestep: 86/1000			Timestep: 87/1000			Timestep: 88/1000			Timestep: 89/1000			Timestep: 90/1000			Timestep: 91/1000			Timestep: 92/1000			Timestep: 93/1000			Timestep: 94/1000			Timestep: 95/1000			Timestep: 96/1000			Timestep: 97/1000			Timestep: 98/1000			Timestep: 99/1000			Timestep: 100/1000			Timestep: 101/1000			Timestep: 102/1000			Timestep: 103/1000			Timestep: 104/1000			Timestep: 105/1000			Timestep: 106/1000			Timestep: 107/1000			Timestep: 108/1000			Timestep: 109/1000			Timestep: 110/1000			Timestep: 111/1000			Timestep: 112/1000			Timestep: 113/1000			Timestep: 114/1000			Timestep: 115/1000			Timestep: 116/1000			Timestep: 117/1000			Timestep: 118/1000			Timestep: 119/1000			Timestep: 120/1000			Timestep: 121/1000			Timestep: 122/1000			Timestep: 123/1000			Timestep: 124/1000			Timestep: 125/1000			Timestep: 126/1000			Timestep: 127/1000			Timestep: 128/1000			Timestep: 129/1000			Timestep: 130/1000			Timestep: 131/1000			Timestep: 132/1000			Timestep: 133/1000			Timestep: 134/1000			Timestep: 135/1000			Timestep: 136/1000			Timestep: 137/1000			Timestep: 138/1000			Timestep: 139/1000			Timestep: 140/1000			Timestep: 141/1000			Timestep: 142/1000			Timestep: 143/1000			Timestep: 144/1000			Timestep: 145/1000			Timestep: 146/1000			Timestep: 147/1000			Timestep: 148/1000			Timestep: 149/1000			Timestep: 150/1000			Timestep: 151/1000			Timestep: 152/1000			Timestep: 153/1000			Timestep: 154/1000			Timestep: 155/1000			Timestep: 156/1000			Timestep: 157/1000			Timestep: 158/1000			Timestep: 159/1000			Timestep: 160/1000			Timestep: 161/1000			Timestep: 162/1000			Timestep: 163/1000			Timestep: 164/1000			Timestep: 165/1000			Timestep: 166/1000			Timestep: 167/1000			Timestep: 168/1000			Timestep: 169/1000			Timestep: 170/1000			Timestep: 171/1000			Timestep: 172/1000			Timestep: 173/1000			Timestep: 174/1000			Timestep: 175/1000			Timestep: 176/1000			Timestep: 177/1000			Timestep: 178/1000			Timestep: 179/1000			Timestep: 180/1000			Timestep: 181/1000			Timestep: 182/1000			Timestep: 183/1000			Timestep: 184/1000			Timestep: 185/1000			Timestep: 186/1000			Timestep: 187/1000			Timestep: 188/1000			Timestep: 189/1000			Timestep: 190/1000			Timestep: 191/1000			Timestep: 192/1000			Timestep: 193/1000			Timestep: 194/1000			Timestep: 195/1000			Timestep: 196/1000			Timestep: 197/1000			Timestep: 198/1000			Timestep: 199/1000			Timestep: 200/1000			Timestep: 201/1000			Timestep: 202/1000			Timestep: 203/1000			Timestep: 204/1000			Timestep: 205/1000			Timestep: 206/1000			Timestep: 207/1000			Timestep: 208/1000			Timestep: 209/1000			Timestep: 210/1000			Timestep: 211/1000			Timestep: 212/1000			Timestep: 213/1000			Timestep: 214/1000			Timestep: 215/1000			Timestep: 216/1000			Timestep: 217/1000			Timestep: 218/1000			Timestep: 219/1000			Timestep: 220/1000			Timestep: 221/1000			Timestep: 222/1000			Timestep: 223/1000			Timestep: 224/1000			Timestep: 225/1000			Timestep: 226/1000			Timestep: 227/1000			Timestep: 228/1000			Timestep: 229/1000			Timestep: 230/1000			Timestep: 231/1000			Timestep: 232/1000			Timestep: 233/1000			Timestep: 234/1000			Timestep: 235/1000-------------------------------------------------------
#> ******** Simulation  7 
#> 			Timestep: 1/1000-------------------------------------------------------
#> ******** Simulation  8 
#> 			Timestep: 1/1000			Timestep: 2/1000			Timestep: 3/1000			Timestep: 4/1000			Timestep: 5/1000			Timestep: 6/1000			Timestep: 7/1000			Timestep: 8/1000			Timestep: 9/1000			Timestep: 10/1000			Timestep: 11/1000			Timestep: 12/1000			Timestep: 13/1000			Timestep: 14/1000			Timestep: 15/1000			Timestep: 16/1000			Timestep: 17/1000			Timestep: 18/1000			Timestep: 19/1000			Timestep: 20/1000			Timestep: 21/1000			Timestep: 22/1000			Timestep: 23/1000			Timestep: 24/1000			Timestep: 25/1000			Timestep: 26/1000			Timestep: 27/1000			Timestep: 28/1000			Timestep: 29/1000			Timestep: 30/1000			Timestep: 31/1000			Timestep: 32/1000			Timestep: 33/1000			Timestep: 34/1000			Timestep: 35/1000			Timestep: 36/1000			Timestep: 37/1000			Timestep: 38/1000			Timestep: 39/1000			Timestep: 40/1000			Timestep: 41/1000			Timestep: 42/1000			Timestep: 43/1000			Timestep: 44/1000			Timestep: 45/1000			Timestep: 46/1000			Timestep: 47/1000			Timestep: 48/1000			Timestep: 49/1000			Timestep: 50/1000			Timestep: 51/1000			Timestep: 52/1000			Timestep: 53/1000			Timestep: 54/1000			Timestep: 55/1000			Timestep: 56/1000			Timestep: 57/1000			Timestep: 58/1000			Timestep: 59/1000			Timestep: 60/1000			Timestep: 61/1000			Timestep: 62/1000			Timestep: 63/1000			Timestep: 64/1000			Timestep: 65/1000			Timestep: 66/1000			Timestep: 67/1000			Timestep: 68/1000			Timestep: 69/1000			Timestep: 70/1000			Timestep: 71/1000			Timestep: 72/1000			Timestep: 73/1000			Timestep: 74/1000			Timestep: 75/1000			Timestep: 76/1000			Timestep: 77/1000			Timestep: 78/1000			Timestep: 79/1000			Timestep: 80/1000			Timestep: 81/1000			Timestep: 82/1000			Timestep: 83/1000			Timestep: 84/1000			Timestep: 85/1000			Timestep: 86/1000			Timestep: 87/1000			Timestep: 88/1000			Timestep: 89/1000			Timestep: 90/1000			Timestep: 91/1000			Timestep: 92/1000			Timestep: 93/1000			Timestep: 94/1000			Timestep: 95/1000			Timestep: 96/1000			Timestep: 97/1000			Timestep: 98/1000			Timestep: 99/1000			Timestep: 100/1000			Timestep: 101/1000			Timestep: 102/1000			Timestep: 103/1000			Timestep: 104/1000			Timestep: 105/1000			Timestep: 106/1000			Timestep: 107/1000			Timestep: 108/1000			Timestep: 109/1000			Timestep: 110/1000			Timestep: 111/1000			Timestep: 112/1000			Timestep: 113/1000			Timestep: 114/1000			Timestep: 115/1000			Timestep: 116/1000			Timestep: 117/1000			Timestep: 118/1000			Timestep: 119/1000			Timestep: 120/1000			Timestep: 121/1000			Timestep: 122/1000			Timestep: 123/1000			Timestep: 124/1000			Timestep: 125/1000			Timestep: 126/1000			Timestep: 127/1000			Timestep: 128/1000			Timestep: 129/1000			Timestep: 130/1000			Timestep: 131/1000			Timestep: 132/1000			Timestep: 133/1000			Timestep: 134/1000			Timestep: 135/1000			Timestep: 136/1000			Timestep: 137/1000			Timestep: 138/1000			Timestep: 139/1000			Timestep: 140/1000			Timestep: 141/1000			Timestep: 142/1000			Timestep: 143/1000			Timestep: 144/1000			Timestep: 145/1000			Timestep: 146/1000			Timestep: 147/1000			Timestep: 148/1000			Timestep: 149/1000			Timestep: 150/1000			Timestep: 151/1000			Timestep: 152/1000			Timestep: 153/1000			Timestep: 154/1000			Timestep: 155/1000			Timestep: 156/1000			Timestep: 157/1000			Timestep: 158/1000			Timestep: 159/1000			Timestep: 160/1000			Timestep: 161/1000			Timestep: 162/1000			Timestep: 163/1000			Timestep: 164/1000			Timestep: 165/1000			Timestep: 166/1000			Timestep: 167/1000			Timestep: 168/1000			Timestep: 169/1000			Timestep: 170/1000			Timestep: 171/1000			Timestep: 172/1000			Timestep: 173/1000			Timestep: 174/1000			Timestep: 175/1000			Timestep: 176/1000			Timestep: 177/1000			Timestep: 178/1000			Timestep: 179/1000			Timestep: 180/1000			Timestep: 181/1000			Timestep: 182/1000			Timestep: 183/1000			Timestep: 184/1000			Timestep: 185/1000			Timestep: 186/1000			Timestep: 187/1000			Timestep: 188/1000			Timestep: 189/1000			Timestep: 190/1000			Timestep: 191/1000			Timestep: 192/1000			Timestep: 193/1000			Timestep: 194/1000			Timestep: 195/1000			Timestep: 196/1000			Timestep: 197/1000			Timestep: 198/1000			Timestep: 199/1000			Timestep: 200/1000			Timestep: 201/1000			Timestep: 202/1000			Timestep: 203/1000			Timestep: 204/1000			Timestep: 205/1000			Timestep: 206/1000			Timestep: 207/1000			Timestep: 208/1000			Timestep: 209/1000			Timestep: 210/1000			Timestep: 211/1000			Timestep: 212/1000			Timestep: 213/1000			Timestep: 214/1000			Timestep: 215/1000			Timestep: 216/1000			Timestep: 217/1000			Timestep: 218/1000			Timestep: 219/1000			Timestep: 220/1000			Timestep: 221/1000			Timestep: 222/1000			Timestep: 223/1000			Timestep: 224/1000			Timestep: 225/1000			Timestep: 226/1000			Timestep: 227/1000			Timestep: 228/1000			Timestep: 229/1000			Timestep: 230/1000			Timestep: 231/1000			Timestep: 232/1000			Timestep: 233/1000			Timestep: 234/1000			Timestep: 235/1000			Timestep: 236/1000			Timestep: 237/1000			Timestep: 238/1000			Timestep: 239/1000			Timestep: 240/1000			Timestep: 241/1000			Timestep: 242/1000			Timestep: 243/1000			Timestep: 244/1000			Timestep: 245/1000			Timestep: 246/1000			Timestep: 247/1000			Timestep: 248/1000			Timestep: 249/1000			Timestep: 250/1000			Timestep: 251/1000-------------------------------------------------------
#> ******** Simulation  9 
#> 			Timestep: 1/1000			Timestep: 2/1000			Timestep: 3/1000-------------------------------------------------------
#> ******** Simulation  10 
#> 			Timestep: 1/1000